dreamjobs/requirements.txt
Django==2.0.2
django-cors-headers==2.2.0
djangorestframework==3.8.2
mysqlclient==1.3.12
Pillow==5.0.0
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
